first draft

+++ crone/src/BufDiskWorker.h
@@ -28,7 +28,7 @@ namespace crone {
     class BufDiskWorker {
     public:
         typedef std::function<void(float secPerSample, float start, size_t count, float* samples)> RenderCallback;
-        typedef std::function<float(int sampleIndex, float inputSample)> ProcessFunc;
+        typedef std::function<float(size_t sampleIndex, float inputSample)> ProcessFunc;
         typedef std::function<void(int jobType)> DoneCallback;
 
     private:
